更新：增加ItemInformationProvider测试。
修复一些旧代码中的bug（虽然已经不再使用）
修复一些byte类型的问题

+++ src/Application.Core/Game/Players/Item.cs
@@ -600,7 +600,7 @@ namespace Application.Core.Game.Players
                             {
                                 if (!(isGM() && YamlConfig.config.server.USE_PERFECT_GM_SCROLL))
                                 {
-                                    it.setUpgradeSlots((byte)(it.getUpgradeSlots() + 1));
+                                    it.setUpgradeSlots(it.getUpgradeSlots() + 1);
                                 }
                                 item = ItemInformationProvider.getInstance().scrollEquipWithId(item, ItemId.CHAOS_SCROll_60, true, ItemId.CHAOS_SCROll_60, isGM());
                             }
